Fix failing tests for QuestionnaireViewModel
