+++ android/quest/src/test/java/org/smartregister/fhircore/quest/medintel/speech/models/GemmaModelTest.kt
@@ -14,6 +14,24 @@
  * limitations under the License.
  */
 
+package org.smartregister.fhircore.quest.medintel.speech.models
+
+/*
+ * Copyright 2021-2024 Ona Systems, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import android.content.Context
 import com.google.mediapipe.tasks.genai.llminference.LlmInference
 import junit.framework.TestCase.assertEquals
